+++ src/Type/AbstractType.php
@@ -146,7 +146,8 @@ abstract class AbstractType implements TypeInterface
             return null;
         }
 
-        $element = Element::create('label[for="' . $this->getIdAttribute() . '"].form-label')->setInnerText($this->label);
+        $element = Element::create('label.form-label')->setInnerText($this->label);
+        $element->attributes['for'] = $this->getIdAttribute();
 
         if ($this->required && !$this->parent) {
             $element->addChild(Element::create('abbr[title="Required Field"]')->setInnerText('*'));